shifted new utils functions to new file
